requirements.txt
matplotlib
networkx
numpy
pyqt